Split the editor hook into focused modules and stabilise parse-dependent tests
Final group from the editor audit (SPECs/editor-audit-spec.md, group 5):

- use-prosemark-editor.ts (860 -> ~215 lines) now only mounts, swaps, and
  disposes the view. Extension assembly moves to editor-extensions.ts;
  search wiring (hidden panel, ancestor-scroller scrollHandler, Mod-f/Mod-g/
  Escape) to editor-search-extensions.ts; link following to
  link-navigation.ts; image/frontmatter paste to editor-clipboard.ts; the
  right-click menu to editor-body-menu.ts; viewport force-parse to
  viewport-parse.ts. No behavior change
- tests/helpers/parsed-state.ts: withFullParse() parses the whole document
  and commits the tree with an empty transaction, so fold/hide/list fields
  are built against it. EditorState.create's wall-clock parse budget could
  expire under a loaded runner, which is what made the mermaid and fold
  fixture tests intermittently fail
- docs/editor.md file map, TODOS, CHANGELOG, and the audit spec status updated
