+++ TODOS.md
@@ -2,11 +2,11 @@
 
 ## In Progress
 
-- Editor audit follow-ups: [`SPECs/editor-audit-spec.md`](SPECs/editor-audit-spec.md) — confirmed bugs (titled links leak their title when folded, list prefix guard/Backspace fire inside code fences, tree-derived ViewPlugins skip parse-advance rebuilds, duplicated stale `foldTreeSync`), keystroke-path performance (full-doc stats/headings per keystroke, whole-tree fold/hide walks, heading guard force-parsing the whole doc), and consolidation of duplicated helpers/command registries. Ordered into five PRs in the spec.
 - Reveal-in-sidebar + residual external-watcher misses: [`SPECs/reveal-in-sidebar-and-external-watcher-spec.md`](SPECs/reveal-in-sidebar-and-external-watcher-spec.md) — keep the explicit tab-context-menu "Reveal in sidebar" action working, leave ordinary file opens from expanding the Everything tree, and characterize the remaining external-file-watcher miss cases through a logging + manual-repro pass before patching further.
 
 ## Done
 
+- Editor audit fixes: [`SPECs/editor-audit-spec.md`](SPECs/editor-audit-spec.md) — five commits: stale-decoration and titled-link bugs plus helper dedupe; keystroke-path performance (deferred stats/headings, indexed fold specs, line-scoped heading guard, cached HTML sanitising, facet-based image src); tree-gated single list-prefix grammar; trimmed basic setup, Escape-closes-find, paste notice, one command registry; hook split into focused modules.
 - Table column sizing (Phase A, CSS-only): [`SPECs/table-column-sizing-spec.md`](SPECs/table-column-sizing-spec.md) — folded table cells inherited `overflow-wrap: anywhere` from `EditorView.lineWrapping`, which feeds min-content sizing and let the auto table layout starve columns until words broke mid-word. Reset to `overflow-wrap: break-word` + `word-break: normal`, replaced the blanket `min-width: 6em` with a `max-width: 48ch` per-cell demand cap, top-aligned cells, and left-aligned headers while keeping explicit `:---:` / `---:` alignment. Phase B (wide tables breaking out of the editor measure) and Phase C (JS column-width computation) remain open.
 - PR #111 review fixes: [`SPECs/Agent/worksheet-pr111-review-fixes.md`](SPECs/Agent/worksheet-pr111-review-fixes.md) — guard creation and external launches against workspace switches, closes, and replaced roots; preserve cross-window global settings writes; and route the sidebar background menu through the full surface and rootless shell.
 - Folder-row Open in Terminal: [`SPECs/sidebar-empty-area-context-menu-spec.md`](SPECs/sidebar-empty-area-context-menu-spec.md) — add an Open in Terminal action to folder row context menus, launching the selected in-workspace directory with the configured terminal.
